requirements.txt
torch==2.4.0
torchvision==0.19.0
tensorflow==2.17.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
opencv-python==4.10.0.84
nibabel==5.2.1
SimpleITK==2.3.1
albumentations==1.4.12
segmentation-models-pytorch==0.3.4
efficientnet-pytorch==0.7.1
grad-cam==1.5.0
matplotlib==3.9.1
seaborn==0.13.2
plotly==5.23.0
tqdm==4.66.4
tensorboard==2.17.0
mlflow==2.15.1
Pillow==10.4.0
scipy==1.14.0
scikit-image==0.24.0
pydicom==2.4.4
reportlab==4.2.2
fastapi==0.111.1
uvicorn==0.30.5
pydantic==2.8.2
onnx==1.16.1
onnxruntime==1.18.1
monai==1.3.2
pytorch-lightning==2.4.0
wandb==0.17.5